fix: sample in tailwind class
ANIMA .
