+++ src/converters/tailwind.ts
@@ -42,6 +42,27 @@ export class TailwindConverter implements IConverter {
     return dsTokens;
   }
 
+  sampleConfigFile(): string {
+    return `
+const {
+  getTwColorsTheme,
+} = require("anima-storybook-cli/dist/lib/getTwColorsTheme");
+const dsToken = require("./design-tokens.json");
+
+const theme = getTwColorsTheme(dsToken);
+
+module.exports = {
+  content: [
+    "./index.html",
+    "./src/**/*.{js,ts,jsx,tsx}",
+  ],
+  theme: {
+    colors: theme,
+  },
+};
+`;
+  }
+
   static convertDSColorToTheme(
     dsTokens: DSTokenTheme,
   ): TailwindConfig['theme']['colors'] {
